+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 Changes:
 
+  * Bulk adding QCs now sorted by alias
   * Fixed V1000 migration hanging on plain sample databases
 
 # 1.4.0